backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
opencv-python-headless>=4.8.0
mediapipe>=0.10.30
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
google-genai>=1.0.0
Pillow>=10.0.0
